LSP Build refactor
Using dependencies instead of bundling for the lsp's actual dependencies
Added @danielx/civet and TypeScript to dependencies
Using published @danielx/civet rather than linked
Sped up build by removing ts-node from build step
Removed CoffeeScript support (moved to plugin)
Greatly reduced bundle size by removing CoffeeScript (it pulled in babel which was > 5MB)
Struggled to improve breakpoints from VS Code (still doesn't work but connecting Dev Tools debugger does)
