+++ lsp/.gitignore
@@ -1,2 +1,3 @@
+*.metafile
 *.vsix
 open-vsx-token
