+++ lsp/.vscode/launch.json
@@ -1,6 +1,7 @@
 {
   "version": "0.2.0",
   "configurations": [
+
     {
       "args": [
         "--extensionDevelopmentPath=${workspaceFolder}"
