+++ lsp/NOTES.md
@@ -32,3 +32,15 @@ In VSCode "Run and Debug" --> "Launch Extension"
 Then in the `chrome://inspect` page you should see: `Remote Target #LOCALHOST` with `/home/daniel/apps/civet/lsp/dist/server.js` and a link to "inspect"
 
 When the DevTools page is open `debugger` statements will work.
+
+Extension Bundle Size
+---
+
+Use [esbuild Bundle Size Analyzer](https://esbuild.github.io/analyze/)
+
+Drop in either:
+
+- `server.metafile`
+- `extension.metafile`
+
+To se what was bundled and why.
