+++ lsp/package.json
@@ -16,7 +16,6 @@
   },
   "activationEvents": [
     "onLanguage:civet",
-    "onLanguage:coffeescript",
     "onLanguage:javascript",
     "onLanguage:javascriptreact",
     "onLanguage:typescript",
@@ -76,7 +75,7 @@
     ],
     "loader": [
       "ts-node/esm",
-      "../dist/esm.mjs"
+      "@danielx/civet/esm"
     ],
     "reporter": "spec",
     "recursive": true,
@@ -91,25 +90,25 @@
     "test": "mocha",
     "watch": "node build.mjs --watch"
   },
+  "dependencies": {
+    "@danielx/civet": "0.6.38",
+    "typescript": "^5.2.2",
+    "vscode-languageclient": "^8.0.2",
+    "vscode-languageserver": "^8.0.2",
+    "vscode-languageserver-textdocument": "^1.0.7"
+  },
   "devDependencies": {
-    "@danielx/civet": "link:..",
     "@danielx/hera": "^0.7.12",
-    "@types/coffeescript": "^2.5.2",
     "@types/mocha": "^9",
     "@types/node": "^14.17.0",
     "@types/vscode": "^1.63.0",
     "@vscode/test-electron": "^2.1.2",
     "@vscode/vsce": "^2.19.0",
-    "coffeescript": "^2.7.0",
     "esbuild": "^0.19.2",
     "esbuild-visualizer": "^0.3.1",
     "mocha": "^10",
     "nyc": "^15.1.0",
     "source-map-support": "^0.5.21",
-    "ts-node": "^10.6.0",
-    "typescript": "^5.2.2",
-    "vscode-languageclient": "^8.0.2",
-    "vscode-languageserver": "^8.0.2",
-    "vscode-languageserver-textdocument": "^1.0.7"
+    "ts-node": "^10.6.0"
   }
 }
