+++ lsp/test/util.civet
@@ -2,7 +2,6 @@
 { intersectRanges, containsRange, makeRange, remapPosition, forwardMap, convertCoffeeScriptSourceMap } from ../source/lib/util.mjs
 assert from assert
 Civet from @danielx/civet
-{compile as coffeeCompile} from coffeescript
 
 describe "util", ->
   it "should intersect ranges", ->
@@ -79,59 +78,3 @@ describe "util", ->
 
     [0, 1, 2, 3, 4, 5, 6, 7].forEach (i) ->
       forwardMap(sourceMap.data.lines, {line: 20 + i, character: 5})
-
-  it "should convert CoffeeScript source maps", ->
-    src := """
-      {parse} = require "./parser"
-      {prune} = gen = require "./generate"
-      {SourceMap} = util = require "./util"
-
-      defaultOptions = {}
-
-      module.exports =
-        parse: parse
-        compile: (src, options=defaultOptions) ->
-          ast = prune parse(src, {
-            filename: options.filename
-          })
-
-          if options.sourceMap
-            sm = SourceMap(src)
-            options.updateSourceMap = sm.updateSourceMap
-            code = gen ast, options
-            return {
-              code,
-              sourceMap: sm
-            }
-
-          gen ast, options
-        generate: gen
-        util: util
-    """
-
-    {js, sourceMap} := coffeeCompile src,
-      bare: true
-      filename: "test.coffee"
-      header: false
-      sourceMap: true
-
-    convertedMap := convertCoffeeScriptSourceMap(sourceMap)
-    {sourceMap: sm2} := Civet.compile "x",
-      sourceMap: true
-
-    // Hack to generate a source map json for testing
-    sm2.data.lines = convertedMap
-    srcMapJSON := sm2.json("yo.coffee", "yo.js")
-    //@ts-ignore
-    srcMapJSON.sourcesContent = [src]
-
-    // console.log src, code
-
-    // console.dir sm.data,
-    //   depth: 8
-
-    base64Encode := (src) ->
-      return Buffer.from(src).toString('base64')
-
-    // Use this to log the code + sourcemap which can be pasted into https://evanw.github.io/source-map-visualization/ see how it works
-    // console.log `${js}\n//# sourceMappingURL=data:application/json;base64,${base64Encode JSON.stringify(srcMapJSON)}`
